Added the ability to process logs incrementally.
That is, it will combine previously persisted data with a new data set.
It works in the following way: if a data set is persisted with
--keep-db-files, then using --load-from-disk will load the original data
set. If new data is passed (piped or through a log file), it will append
it to the original data set. To preserve the data at all times,
--keep-db-files must be used. If --load-from-disk is used without
--keep-db-files, database files will be deleted upon closing the
program.