Added command-line shortcuts for standard log formats.
Now it allows to use existing log/date/time format options such as:
--log-format=COMBINED which will automatically set the log/date/time
formats.

If a --date-format or --time-format is supplied, it may overwrite
(depending on which option is supplied last) whatever option was set
with --log-format. For instance,

  --log-format=COMBINED --date-format=W3C

will make use of the combined log/time format and the w3c date format.