Ensure --http-method/protocol have the proper default value set.
Options that can be toggled, require an argument, either <yes|no>.

On the other hand, options that don't require an argument can be used
simply by setting the option in the command line, i.e., --agent-list.
Additionally, these options can have a value of 'true' in the config
file which will enable the option, or false and the config file parser
will ignore the option.