Implemented a throttle mechanism for slow clients.
Some client's are not able to keep up with the rate at which data is
being transferred from the server. So if a client hits the threshold
of 2MiB, the server won't keep queuing data until everything in the
queue is sent. This fixes the issue where the server could buffer too
much data while slow clients are trying to keep up.