Fixed issue where parser couldn't read some X-Forwarded-For formats.
This changes how XFF is currently parsed so goaccess can support
multiple XFF formats (see issue #632 and #615).

This breaks compatibility with the original implementation of parsing
XFF. It now makes use of a special specifier which consists of a tilde
before the host specifier, followed by the character(s) that delimit
the XFF field, which are enclosed by curly braces (i.e., ~h{," }).

For instance, `~h{," }` is used in order to parse the following XFF
field which is delimited by a double quote, a comma, and a space.

"11.25.11.53, 17.68.33.17"

In the case above, the parser will ignore any of the character(s)
within the curly braces and attempt to extract the first valid IPv4/6.