websocket: create ws_fifo in ws_init not ws_start
Move the creation of the websocket fifo from ws_start() to ws_init().

There exists a race condition where the websocket thread gets delayed
until after the main thread runs.  The main thread then tries to open
the fifo before the websocket has a chance to create the fifo,
resulting in a crash.

By moving this to ws_init(), it can be called from the main thread.

Signed-off-by: Sean Cross <sean@xobs.io>