Enable out-of-tree builds
autoconf/automake usually allows a user to do an out-of-tree build:

    mkdir build
    cd build
    ../configure
    make

This is convenient to easily start from scratch by just removing the
`build` directory (whatever we are working from a git checkout or a
tarball).

To make it work, manually written rules, like these used to embed
HTML/JS/CSS resources, need to be rewritten to correctly address
source files.