Add an option to disable client IP validation
To avoid logging IP addresses, users may obfuscate them using a
regularly rotated seed and an hash algorithm. For nginx, [ipscrub][]
can be used for such a task. However, the result is not a valid IP
address.

This commit adds an option, `--no-ip-validation`, to avoid checking
the validity of an IP address and use the provided field as is. The IP
is stored as a string and the current code correctly handles the case
where this field is not an IP address (`inet_pton()` is checked with
`AF_INET`, then `AF_INET6` and falls back to some sensible behaviour
for other cases).

Fix #1354

[ipscrub]: https://github.com/masonicboom/ipscrub