Allow `get_home ()` to return NULL on error, instead of terminating (Closes: #1849)
When running from a systemd service, there's very few environment
variables defined. Which env variables are available exactly is
explained at [1] (long read).

Short version, here's what `printenv` gives me, in a script that is
spawned by logrotate, which is itself started daily by systemd:

    + printenv
    PWD=/
    LANG=en_US.UTF-8
    INVOCATION_ID=dbefda3a93634d16bca7ba0e077d57da
    SHLVL=1
    JOURNAL_STREAM=9:52249
    PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
    _=/usr/bin/printenv

As one can see, `HOME` is not defined, and because of that GoAccess
fails:

    GoAccess - version 1.4 - May 18 2020 01:45:26
    Config file: No config file used
    Fatal error has occurred
    Error occurred at: src/util.c - get_home - 453
    Unable to determine the HOME environment variable.

This commit changes GoAccess behavior a bit, and makes it keep running
if it can't get the value of HOME, instead of hitting a FATAL() and
exiting.

The only caller of `get_home ()` is `get_config_file_path ()`, and it
already supports the case where `get_home ()` returns NULL. It will drop
a debug log message and keep going:

    LOG_DEBUG (("Unable to find user's config file %s %s", upath, strerror (errno)));

----

[1]: https://www.freedesktop.org/software/systemd/man/systemd.exec.html#Environment%20variables%20in%20spawned%20processes