Fixed issue where log truncation could prevent accurate number counting.
There are two scenarios that got fixed:

1. The log got smaller, probably was truncated so we start reading from 0,
before it was reading from the last log known size.

2. If the log got larger since the last 'tail' iteration, we attempt to compare
the first READ_BYTES against the READ_BYTES we had since the last parse. If
it's different, then it means the file may got truncated but grew faster than
the last iteration (odd, but possible), so we read from 0.

This is not ideal as we need to check the first READ_BYTES each tail iteration,
but maybe the only reliable way to know if the  current log looks different
than our first read/parse.