POSIX has made usleep(2) obsolete in favour of nanosleep(2)
usleep(2) also has no effect if equal to greater than 1000000
which happens to match HTML_REFERS, but no error was checked.
This caused 100% CPU utiliation constantly following the tail
of the web server access log with default value.

nanosleep(2) has more relaxed limits.
This commit does a like for like replacement and also adds
error checking in case of a similar error in the future.

Signed-off-by: Roy Marples <roy@marples.name>