Replace select(2) with poll(2)
poll(2) is more modern, more efficient and a lot faster than
select(2). The trade-off is that we need to manage the
array of struct pollfd we send it a bit more intelligently
than just manipulating tables before each call to select(2).

This commit adds functions to dynamically resize the allocation
needed when new fd's need to be monitored or old ones discarded
and then sprinkles them where needed.