Monitor log files using kqueue(2)
This patch replaces the behaviour of opening, checking,
processing and closing each logfile every few moments with a
kqueue(2) based file monitor.

The new behaviour is similar to poll(2) in that the OS will inform us
when the file has changed. To do this, we need to hold a descriptor
open at all times rather than close it between runs.
If the logfile is turned over, we then monitor the directory for
any modifications in the same way and on a change we try to re-open
the logfile once more.

This slightly reduces the CPU time goaccess spends on NetBSD at least,
but is a massive saving on the CPU time the kernel spends constantly
stating, opening and closing the logfile.