Drop unnecessary NULL check
The variable `key` is checked in the beginning of the function to be
non-NULL.
In the while loop it gets modified via `key++;` making it impossible to
be NULL after the loop.

check_after_deref: Null-checking key suggests that it may be null, but it has already been dereferenced on all paths leading to the check.

Found by Coverity