Fixed regression introduced on ae98eb9 when parsing UA.
See:
https://github.com/allinurl/goaccess/issues/1573
https://github.com/allinurl/goaccess/pull/2266

It now moves the replacing of '+' with ' ' right when checking the browser or
OS instead of replacing the original UA. This also affects any char
replacements on the referrers and requests. However, in theory those shouldn't
need '+'s to be replaced as the decoder will take care of that work (if
needed).