Fixed issue parsing a delimited XFF field followed by a host IP.
This also changes the way the XFF specifier `~h` uses its delimiters.  Prior to
this commit, ~h would require that all delimiters needed to be inside the
braces. For instance:

"192.1.2.13, 192.68.33.17, 192.1.1.2" => ~h{", }
"192.1.2.12", "192.68.33.17"          => ~h{", }
192.1.2.12, 192.68.33.17              => ~h{, }
192.1.2.14 192.68.33.17 192.1.1.2     => ~h{ }

Now, it only requires those that are within the XFF field. So for the above
examples, it works now as:

"192.1.2.13, 192.68.33.17, 192.1.1.2" => "~h{, }"
"192.1.2.12", "192.68.33.17"          => ~h{", }
192.1.2.12, 192.68.33.17              => ~h{, }
192.1.2.14 192.68.33.17 192.1.1.2     => ~h{ }