Fix null pointer exception on FreeBSD (fixes #2476)
On FreeBSD, goaccess dumps core when reading from STDIN.

FreeBSD's basename(3) behaves a little different from the rest
of the world. Rather than sometimes returning a pointer to a new
allocation, it always returns a pointer into the buffer itself.

When basename() gets a string, including an empty string, it
behaves as expected. But if it's given a null pointer, it
returns another null pointer rather than a pointer to a
newly-allocated empty string. This makes it thread-safe, but
(as usual for FreeBSD) assumes that the rest of the world
enjoys FreeBSD's quirky null pointer exceptions.

When reading from STDIN, the filenames vector is empty, so
set_glog's filename arg is empty. Calling basename on it makes
a null pointer, and subsequent use causes a segfault.

Calling xstrdup on the filename is a pointless but trivial
duplication for the non-BSD world, but guarantees a string of
some sort for FreeBSD.