Extract referer host more reliably
It is wrong to assume that the host part is always ended by an '/'
character.

When an URL contains URL parameters at the end without a leading '/'
character, then the strchr() search might leak into the content of the
URL parameters.

For example consider the following referer:

https://example.com?ref=https%3A//giter.vip

This is not to be confused with the more common variant with a leading
slash:
https://example.com/?ref=https%3A//giter.vip

With the first variant, the parser will choke on the missing / and
instead mistakenly match the whole "example.com?ref=https%3A" as the
host part, as this is where the first slash appears.

To fix this, we now also accept "?" as the end marker of the host part,
as this is a perfectly valid URL scheme according to the spec[0] - and
indeed it does sometimes appear in logs.

[0] https://url.spec.whatwg.org/#url-writing