Basic first attempt at returning the HTML report with server
If you're serving real-time reports in an ad-hoc manner, it can be
annoying to set up another web server to return the HTML. This makes it
so that the same web server used for WebSocket updates can also serve
the HTML.

Many problems likely; I have not yet audited this. It is somewhat buggy
too - ws_send_response likes to not return anything and thus truncates
the buffer. I'm trying to find a solution to this.

It also probably needs to handle the case when static resources are
served in many different files.